weird fix but ok

+++ proj1/proj1.h
@@ -78,7 +78,7 @@ void remove_comments(String* input, String *output);
 
 // STATE MACHINE FUNCTIONS
 void runtime(MacroList *list, String *input, String *output);
-size_t process_macro(MacroList *list, String *input, String *output, size_t i);
+int process_macro(MacroList *list, String *input, String *output, size_t i);
 size_t find_close_brace(String *input, size_t i);
 
 // EXPAND FUNCTIONS
